DONE: Translation, basic error checking, two languages.
I am basically uploading this a backup. Will clean it up later.

Languages corrected (To the best of my abilities):

- Norwegian
- German

Icelandic is Work in Progress, about 50% done with the manual error correction (Since it isn't a popular language, there's a lot of mistakes in the automatic correction, but most translators don't even recognize it).